# User/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Este API REST se encarga de gestionar la información y operaciones relacionadas con los usuarios dentro del microservicio de Usuarios del sistema de Netflix. Proporciona diversos endpoints para manejar el ciclo de vida completo de los usuarios, incluyendo la creación de cuentas, autenticación, actualización de perfiles, y la gestión de preferencias y configuraciones de cada usuario. Permite realizar operaciones CRUD (Crear, Leer, Actualizar, Eliminar) sobre los datos de usuario, asegurando que la información esté siempre segura y actualizada. Esta API facilita la interacción entre el frontend y el backend del sistema, garantizando que los usuarios puedan gestionar fácilmente sus cuentas y preferencias, mejorando la experiencia personalizada y segura dentro de la plataforma.
  title: API de Usuario
  version: "1.0"
servers:
- url: https://api.server/v1/user/
...
paths:
...
  /user/{userID}/list/{listID}/film/{filmID}:
    delete:
      description: Permite elimnar una pelicula de una lista
      parameters:
      - description: ID del usuario que posee la lista
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la lista a la que se desea añadir contenido
        explode: false
        in: path
        name: listID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la pelicula
        explode: false
        in: path
        name: filmID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "201":
          description: Pelicula eliminada de la lista
        "400":
          description: Solicitud inválida (datos incompletos o incorrectos)
        "404":
          description: Lista no encontrada
        "409":
          description: Conflicto (no se puede añadir el contenido a la lista)
      summary: Elimina una pelicula de una lista
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
    post:
      description: Permite agregar nuevas peliculas a una lista específica.
      parameters:
      - description: ID del usuario que posee la lista
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la lista a la que se desea añadir contenido
        explode: false
        in: path
        name: listID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la pelicula
        explode: false
        in: path
        name: filmID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "201":
          description: Contenido añadido a la lista con éxito
        "400":
          description: Solicitud inválida (datos incompletos o incorrectos)
        "404":
          description: Lista no encontrada
        "409":
          description: Conflicto (no se puede añadir el contenido a la lista)
      summary: Añade pelicula a una lista existente
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
  /user/{userID}/list/{listID}/serie/{serieID}:
    delete:
      description: Permite eliminar una serie de una lista
      parameters:
      - description: ID del usuario que posee la lista
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la lista a la que se desea añadir contenido
        explode: false
        in: path
        name: listID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la serie
        explode: false
        in: path
        name: serieID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "201":
          description: Serie eliminada de la lista
        "400":
          description: Solicitud inválida (datos incompletos o incorrectos)
        "404":
          description: Lista no encontrada
        "409":
          description: Conflicto (no se puede añadir el contenido a la lista)
      summary: Elimina una serie de una lista
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
    post:
      description: Permite agregar nuevas series a una lista específica.
      parameters:
      - description: ID del usuario que posee la lista
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la lista a la que se desea añadir contenido
        explode: false
        in: path
        name: listID
        required: true
        schema:
          type: integer
        style: simple
      - description: ID de la serie
        explode: false
        in: path
        name: serieID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "201":
          description: Contenido añadido a la lista con éxito
        "400":
          description: Solicitud inválida (datos incompletos o incorrectos)
        "404":
          description: Lista no encontrada
        "409":
          description: Conflicto (no se puede añadir el contenido a la lista)
      summary: Añade serie a una lista existente
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
...